[TM-1272] Rename hook.
