--- src/hooks/usePrevious.ts
@@ -1,19 +1,0 @@
-import { useEffect, useRef } from "react";
-
-export function usePrevious<T>(value: T): T | undefined {
-  const ref = useRef<T>();
-  useEffect(() => {
-    ref.current = value;
-  }, [value]);
-  return ref.current;
-}
-
-export function useValueChanged<T>(value: T, callback: () => void) {
-  const ref = useRef<T>();
-  useEffect(() => {
-    if (ref.current !== value) {
-      ref.current = value;
-      callback();
-    }
-  });
-}
